+++ CHANGELOG.md
@@ -17,6 +17,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Shortest paths storing
 - Possibility to provide a list of init input files
 - PT delay information factory
+- Added an environment variable OUTPUT_FOLDER to specify the output folder
 
 ### Changed
 - Deprecated use of origin/destination_lat/lon, now fetching coordinates from geometry